#16 Move projects to Global scope
