+++ my-pwa/src/components/Pages/Search.js
@@ -14,7 +14,6 @@ import Divider from 'material-ui/Divider';
 import TextField from 'material-ui/TextField';
 
 import * as globalActions from '../../actions/global'
-import * as projectActions from '../../actions/projects'
 
 import {
   isEmpty,
@@ -160,6 +159,5 @@ export default connect(
   }),
   (dispatch) => ({
     actions: bindActionCreators(globalActions, dispatch),
-    projects: bindActionCreators(projectActions, dispatch),
   })
 )(Search)
