+++ my-pwa/src/reducers/index.js
@@ -1,12 +1,10 @@
 import { combineReducers } from 'redux';
 import global from './global';
 import tasks from './tasks';
-import projects from './projects';
 
 const appReducer = combineReducers({
   global,
   tasks,
-  projects,
   
 });
 export default appReducer;
